pyproject.toml
[project]
name = "cc-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.56.0",
    "requests>=2.32.5",
]
